reuse dtype_dl2dnnl()
